// packages/datadog-api-client-v1/models/ServiceLevelObjectiveQuery.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* A metric SLI query. **Required if type is `metric`**. Note that Datadog only allows the sum by aggregator to be used because this will sum up all request counts instead of averaging them, or taking the max or min of all of those requests.
*/
export declare class ServiceLevelObjectiveQuery {
    /**
    * A Datadog metric query for total (valid) events.
    */
    'denominator': string;
    /**
    * A Datadog metric query for good events.
    */
    'numerator': string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): ServiceLevelObjectiveQuery;
    static serialize(data: ServiceLevelObjectiveQuery): {
        [key: string]: any;
    };
    constructor();
}